🎨 Plotkin Theorem 1
Add `demo.rkt` showing exactly Plotkin's example
Fix an error in the `substitute` that could cause infinite loops
Fix formatting in inductive-generalization

There was previously an error in the substitution function. If a
Term was used in multiple places in a Word, it was possible for
the Term to be replaced even if the Terms agreed in both words.

For example:

```
W1 = P(a(), b())
W2 = P(a(), a())
```

Should be antiunified as:

```
P(a(), _0)
```

The previous version of substitution would have replaced both
`a()` in W2, causing an infinite loop.